Extract local_repo
